requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2
langgraph>=0.2.0
langchain-core>=0.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
chromadb==0.4.22
numpy==1.26.4
pydantic==2.8.2
